+++ server/src/main/java/com/moovda_project/moovda/module/movie/api/MovieApiService.java
@@ -41,8 +41,8 @@ public class MovieApiService {
 
                 Movie movie = new Movie();
 
-                String posterUrl = movieObj.get("posters").toString().split("\\|")[0];
-                if(posterUrl.equals("")) continue;        // 포스터는 하나만 저장
+                String posterUrl = movieObj.get("posters").toString().split("\\|")[0];  // 포스터는 하나만 저장
+                if(posterUrl.equals("")) continue;       // 포스터가 없으면 저장하지 않는다.
 
                 movie.setPoster(posterUrl);
 
@@ -114,7 +114,6 @@ public class MovieApiService {
 
                     if(directorCnt+actorCnt==8) break;  // 최대 8명까지만 저장
                 }
-
                 // 스태프 끝
 
                 movieRepository.save(movie);
