+++ server/src/main/java/com/moovda_project/moovda/module/movie/controller/MovieController.java
@@ -1,7 +1,7 @@
 package com.moovda_project.moovda.module.movie.controller;
 
 import com.moovda_project.moovda.global.dto.SingleResponseDto;
-import com.moovda_project.moovda.module.movie.dto.MovieSearchCondition;
+import com.moovda_project.moovda.module.movie.dto.search.MovieSearchCondition;
 import com.moovda_project.moovda.module.movie.entity.Movie;
 import com.moovda_project.moovda.module.movie.mapper.MovieMapper;
 import com.moovda_project.moovda.module.movie.service.MovieService;
@@ -35,7 +35,7 @@ public class MovieController {
                                       @Positive @RequestParam int page) {
           List<Movie> movies = movieService.filterMovie(condition);
 
-          return new ResponseEntity<>(mapper.moviesToMovieFilterResponseDtosV2(movies,page,5), HttpStatus.OK);
+          return new ResponseEntity<>(mapper.moviesToPagedMovieFilterResponseDto(movies,page,5), HttpStatus.OK);
 
     }
 }
