+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/MovieFilterResponseDto.java
@@ -13,5 +13,4 @@ public class MovieFilterResponseDto {
     String poster;
     Double starAvg;
 
-
 }
