+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/PagedMovieFilterResponseDto.java
@@ -1,5 +1,6 @@
 package com.moovda_project.moovda.module.movie.dto;
 
+import com.moovda_project.moovda.global.dto.PageDto;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,9 +8,7 @@ import java.util.List;
 
 @Getter
 @Setter
-public class MovieFilterResponseDtoV2 {
-
+public class PagedMovieFilterResponseDto {
     List<MovieFilterResponseDto> movies;
-
     PageDto pageInfo;
 }
