+++ server/src/main/java/com/moovda_project/moovda/module/movie/dto/search/MovieSearchCondition.java
@@ -1,4 +1,4 @@
-package com.moovda_project.moovda.module.movie.dto;
+package com.moovda_project.moovda.module.movie.dto.search;
 
 
 import lombok.AllArgsConstructor;
