+++ server/src/main/java/com/moovda_project/moovda/module/movie/service/MovieService.java
@@ -3,8 +3,8 @@ package com.moovda_project.moovda.module.movie.service;
 
 import com.moovda_project.moovda.global.exception.BusinessLogicException;
 import com.moovda_project.moovda.global.exception.ExceptionCode;
-import com.moovda_project.moovda.module.movie.dto.MovieSearchCondition;
-import com.moovda_project.moovda.module.movie.dto.MovieSearchDto;
+import com.moovda_project.moovda.module.movie.dto.search.MovieSearchCondition;
+import com.moovda_project.moovda.module.movie.dto.search.MovieSearchDto;
 import com.moovda_project.moovda.module.movie.entity.Movie;
 import com.moovda_project.moovda.module.movie.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +18,6 @@ import java.util.*;
 @Transactional
 @RequiredArgsConstructor
 public class MovieService {
-
     private final MovieRepository movieRepository;
 
     @Transactional(readOnly = true)
@@ -41,6 +40,12 @@ public class MovieService {
         Set<Long> movieIds = new HashSet<>();
         List<Movie> filteredMovies = new ArrayList<>();
 
+        removeSameMovie(movieSearchDtos, movieIds, filteredMovies); // 중복 영화 제거
+
+        return filteredMovies;
+    }
+
+    private void removeSameMovie(List<MovieSearchDto> movieSearchDtos, Set<Long> movieIds, List<Movie> filteredMovies) {
         for (MovieSearchDto movieSearchDto : movieSearchDtos) {
             if (!movieIds.contains(movieSearchDto.getMovieId())) {
                 movieIds.add(movieSearchDto.getMovieId());
@@ -48,8 +53,6 @@ public class MovieService {
                 filteredMovies.add(movie);
             }
         }
-
-        return filteredMovies;
     }
 
     private Movie findverifiedMovie(long movieId) {
